requirements.txt
Django==1.8.5
djangorestframework==3.2.5
Markdown==2.6.5
psycopg2==2.6.1
wheel==0.24.0
